Fixed common library switch
